Testing AI Select and Worker Class Implementation
Atom Beautify all javascript.
Fill in the GameWorld Create methods in preparation using the rock.png as a test.

Implement Worker Class to be able to switch active AI.

Fill in FSM and Utility to be able to become active or stop.
Allow AI to be selected from the AI select screen.
Test FSM and Utility by running a test state or action when selected from the AI Select Screen.
Allow the gameWorld to become active only when the InGame screen is open.
Load Test Assets