Imlement a stamina system into the worker and finite state machine.
Added a stamina bar sprite to the worker and necessary methods to update the sprite width.
Added conditions to the finite state machine AI that enables the "Resting" state when stamina is low.
Modified collision methods and state switch to modify stamina.
Added new Rock and Stone sprites and temporary banners and buttons for UI.
Removed the Text object attached to asset sprites.
Renamed some assets.